requirements.txt
streamlit==1.28.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.0.76
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2